+++ project-rival/Game.h
@@ -45,7 +45,7 @@ public:
 
 protected:
 	void init();
-	void update(double dt);
+	void update(float dt);
 
 	void ManageWave();
 
@@ -61,7 +61,7 @@ protected:
 
 #ifdef TEST_FPS
 	sf::Text x_updateFPS{ m_arialFont };	// text used to display updates per second.
-	sf::Text x_drawFPS{  m_arialFont };	// text used to display draw calls per second.
+	sf::Text x_drawFPS{ m_arialFont };	// text used to display draw calls per second.
 	sf::Time x_secondTime{ sf::Time::Zero };			// counter used to establish when a second has passed.
 	int x_updateFrameCount{ 0 };						// updates per second counter.
 	int x_drawFrameCount{ 0 };							// draws per second counter.
@@ -71,8 +71,8 @@ private:
 	// ----------------------------------> FUNCTIONS <---------------------------------- //
 	// Update subfunctions
 	void CollisionChecks();
-	void gameInput();	
-	
+	void gameInput();
+
 	// Game management
 	void resetGame();
 	void gameStart();
@@ -92,7 +92,7 @@ private:
 	// Game Management
 	sf::View m_gameCamera;
 	bool m_isPlayerCamera;
-	
+
 	// Player management
 	Player m_player;
 	bool m_isInCombat;
